+++ sycl/source/detail/device_info.hpp
