+++ sycl/source/detail/memory_manager.cpp
@@ -7,10 +7,10 @@
 //===----------------------------------------------------------------------===//
 
 #include <CL/sycl/detail/memory_manager.hpp>
-#include <CL/sycl/detail/usm_dispatch.hpp>
 #include <detail/context_impl.hpp>
 #include <detail/event_impl.hpp>
 #include <detail/queue_impl.hpp>
+#include <detail/usm/usm_dispatch.hpp>
 
 #include <algorithm>
 #include <cassert>
