+++ sycl/source/detail/platform_util.hpp
