+++ sycl/source/detail/queue_impl.cpp
@@ -10,9 +10,9 @@
 #include <CL/sycl/detail/clusm.hpp>
 #include <CL/sycl/detail/memory_manager.hpp>
 #include <CL/sycl/detail/pi.hpp>
-#include <CL/sycl/detail/usm_dispatch.hpp>
 #include <CL/sycl/device.hpp>
 #include <detail/queue_impl.hpp>
+#include <detail/usm/usm_dispatch.hpp>
 
 #include <cstring>
