+++ sycl/source/detail/usm/usm_dispatch.cpp
@@ -8,7 +8,7 @@
 
 #include <CL/sycl/detail/pi.hpp>
 #include <CL/sycl/detail/plugin.hpp>
-#include <CL/sycl/detail/usm_dispatch.hpp>
+#include <detail/usm/usm_dispatch.hpp>
 
 __SYCL_INLINE_NAMESPACE(cl) {
   namespace sycl {
